End Sprint #6 (#82)
* changes in appearence and UX for context management, deleted context/[id] page

* context page enhancement

* added name and surname of users to the table

* Add Microsoft Auth.
Must clean code

* Fixed POST request pointing to the wrong API call

* Fixed tenant to be common and not app-registry only

* Add new variable for Microsoft OAUTH

* Create Component for Microsoft OAuth

* Added possibility to Accept an invitation with Microsoft Account

* Now it uses MicrosoftOAuth Component
+ search for state containing the invitation token.

* Now it uses Microsoft OAuth

* added full name of users when relevant

* added rfcs and adrs status on user dashboard

* Updated comments

* Finished chatbot on frontend

* Refactor setLoginSession

* Use refactored method

* Handle errors

* refactor specific rfc and split discussion tab

* diagram creation and edit now happens through a draw.io modal window

* Fixed small bugs

* Resolve all merge conflicts

* Fix New Alternative not using DrawIO component

---------
