+++ .env.example
@@ -1,2 +1,3 @@
 NEXT_PUBLIC_BACKEND_URL=http://localhost:8000
 NEXT_PUBLIC_SITE_URL=http://localhost:3000
+MICROSOFT_OAUTH_CLIENT_ID=""
