+++ src/app/(authed)/dashboard/page.tsx
@@ -5,6 +5,7 @@ import Link from "next/link";
 import { Card } from "@/components/ui/card";
 import type { AdrResponse, RfcResponse } from "@/lib/types";
 import { authFetch } from "@/lib/fetcher";
+import { ChatbotWidget } from "@/components/chatbot/ChatbotWidget";
 
 function formatRelative(iso?: string) {
   if (!iso) return "";
@@ -81,6 +82,38 @@ export default function DashboardPage() {
   const [suggestionsLoading, setSuggestionsLoading] = useState(false);
   const [showSuggestions, setShowSuggestions] = useState(false);
 
+
+
+  const getStatusColorRFC = (status: RfcResponse['status']) => {
+    switch (status) {
+      case 'CLOSED_DECIDED':
+      case 'CLOSED_NON_DECIDED':
+        return 'bg-purple-200 text-purple-800';
+      case 'UNDER_REVIEW':
+        return 'bg-green-200 text-green-800';
+      default:
+        return 'bg-gray-200 text-gray-700';
+    }
+  };
+
+  const formatStatusRFC = (status: RfcResponse['status']) => {
+    return status.replace(/_/g, ' ');
+  };
+
+
+  const getStatusColorADR = (status: AdrResponse['status']) => {
+    switch (status) {
+      case 'APPROVED':
+        return 'bg-green-200 text-green-800';
+      default:
+        return 'bg-gray-200 text-gray-700';
+    }
+  };
+
+  const formatStatusADR = (status: AdrResponse['status']) => {
+    return status.replace(/_/g, ' ');
+  };
+
   // Recent RFCs / ADRs
   useEffect(() => {
     (async () => {
@@ -141,7 +174,7 @@ export default function DashboardPage() {
     (async () => {
       try {
         const res = await authFetch(
-          `${process.env.NEXT_PUBLIC_BACKEND_URL}/users`,
+          `${process.env.NEXT_PUBLIC_BACKEND_URL}/users?page=0&size=100`,
         );
         if (!res.ok) {
           console.error("Failed to load users:", res.status);
@@ -270,7 +303,7 @@ export default function DashboardPage() {
 
         const params = new URLSearchParams();
         params.set("q", q);
-        params.set("sort", sort); 
+        params.set("sort", sort);
         if (authorFilter) params.set("authorId", authorFilter);
         if (dateFrom) params.set("dateFrom", dateFrom);
         if (dateTo) params.set("dateTo", dateTo);
@@ -629,9 +662,15 @@ export default function DashboardPage() {
                   className="flex items-center justify-between py-3 hover:bg-muted/20 transition-colors"
                 >
                   <span className="text-sm font-medium text-foreground text-gray-700">{rfc.title}</span>
-                  <span className="text-sm text-gray-500">{formatRelative(pickTimestamp(rfc))}</span>
+                  <div className="flex items-center gap-2 ml-auto">
+                    <span className="text-xs text-gray-400">{formatRelative(pickTimestamp(rfc))}</span>
+                    <span className={`px-3 py-1 rounded-full text-xs font-medium ${getStatusColorRFC(rfc.status)}`}>
+                      {formatStatusRFC(rfc.status)}
+                    </span>
+                  </div>
                 </Link>
               ))
+
             )}
           </Card>
         )}
@@ -665,13 +704,21 @@ export default function DashboardPage() {
                   className="flex items-center justify-between py-3 hover:bg-muted/20 transition-colors"
                 >
                   <span className="text-sm font-medium text-foreground text-gray-700">{adr.title}</span>
-                  <span className="text-sm text-gray-500">{formatRelative(pickTimestamp(adr))}</span>
+                  <div className="flex items-center gap-2 ml-auto">
+                    <span className="text-xs text-gray-400">{formatRelative(pickTimestamp(adr))}</span>
+                    <span className={`px-3 py-1 rounded-full text-xs font-medium ${getStatusColorADR(adr.status)}`}>
+                      {formatStatusADR(adr.status)}
+                    </span>
+                     </div>
                 </Link>
               ))
             )}
           </Card>
         )}
       </section>
+
+      {/* Floating Chatbot widget */}
+      <ChatbotWidget />
     </main>
   );
 }
