+++ src/components/chatbot/types.ts
@@ -1,0 +1,8 @@
+export type ChatRole = "user" | "assistant";
+
+export type ChatMessage = {
+  id: string;
+  role: ChatRole;
+  content: string;
+  createdAt: number;
+};
